Create a filter component
